Reworked CPR. Moved CPR to dedticated class. CPR API additions.
CPR is now state-based rather than relying on GameFibers. Animations for
both the player and victim are working. Added basic contols for doing CPR
that should be moved to control settings. The number of required pumps to
revive are now dynamic. Removed rhythm and pulse monitors while doing CPR.

API addition: WasCPRSuccessful(Ped) returns a nullable boolean to check if the
CPR attempt was successful.